# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: nom-indexer-go API
  version: 0.1.0
  description: |
    Read-only HTTP API over the nom-indexer-go database. Schema-as-contract:
    every response mirrors the underlying Postgres table layout documented at
    https://0x3639.github.io/nom-indexer-go/schema/.

    ## Authentication

    All endpoints under `/api/v1/` require a Bearer JWT in the `Authorization`
    header. Tokens are HS256-signed by an admin via `cmd/jwt-issue`; there is
    no token-mint endpoint. See https://0x3639.github.io/nom-indexer-go/api/auth/.

    ## Pagination

    Collection endpoints use offset/limit (`?page=1&page_size=50`, max 200) and
    return an envelope `{data: [...], pagination: {page, page_size, total}}`.
    Single-item endpoints return the object directly.

    ## Errors

    Errors are returned as `application/problem+json` (RFC 7807) with a
    JSON object carrying `type`, `title`, `status`, `detail`, and a
    custom `code` field.
  contact:
    name: nom-indexer-go
    url: https://github.com/0x3639/nom-indexer-go
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256-signed JWT minted via `cmd/jwt-issue`. Carries `sub`, `exp`, `iat`,
        and an OAuth 2.0 space-separated `scope` claim.
...
  schemas:
...
    Status:
      type: object
      description: Indexer sync state derived from the database.
      required: [latest_height, latest_timestamp, indexer_lag_seconds, version]
      properties:
        latest_height:
          type: integer
          format: int64
          minimum: 0
          description: Height of the most-recently-indexed momentum (0 if empty).
          examples: [12345]
        latest_timestamp:
          type: integer
          format: int64
          description: Unix-seconds timestamp of the latest momentum.
          examples: [1700000000]
        indexer_lag_seconds:
          type: integer
          format: int64
          description: Server clock minus latest_timestamp. Grows when the indexer falls behind.
          examples: [5]
        version:
          type: string
          examples: ["dev"]
...
paths:
...
  /api/v1/status:
    get:
      operationId: getStatus
      summary: Indexer sync status
      description: |
        Returns the latest indexed momentum height, its timestamp, and
        `indexer_lag_seconds = now - latest_timestamp`. Computed from the
        database alone — does not hit the Zenon node.
      tags: [meta]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current sync state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/momentums:
    get:
      operationId: listMomentums
      summary: List momentums
      description: |
        Returns momentums ordered by height (default `desc`). Page through
        the full chain with `?page=` and `?page_size=`.
      tags: [momentums]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated list of momentums.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MomentumList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/momentums/stream:
    get:
      operationId: streamMomentums
      summary: WebSocket stream of new momentums
      description: |
        Upgrades the connection to a WebSocket and pushes one JSON
        frame per newly-indexed momentum. Frame shape matches the
        `Momentum` schema this spec defines for the REST endpoints.

        ## Auth

        Either of:

        - `Authorization: Bearer <token>` header (CLI, server-side).
        - `?token=<jwt>` query parameter (browsers — the WebSocket
          API cannot send custom headers on the upgrade request).
          Tokens passed this way may end up in HTTP proxy logs;
          prefer short TTLs for stream consumers.

        ## Replay

        Pass `?from_height=N` to backfill momentums starting at
        height `N` before switching to live streaming. The catch-up
        scan is capped at 10,000 rows — beyond that, fall back to
        the REST `/api/v1/momentums` endpoint for the historical
        gap and reconnect for live.

        ## Close codes

        | Code | Meaning |
        |---|---|
        | 1000 | Normal: client disconnected or server shutdown. |
        | 1011 | Internal: replay or dispatch error. |
        | 4000 | `slow_consumer`: the hub dropped frames; reconnect with `from_height` of the last height you saw. |

        Auth failures during the HTTP upgrade come back as
        `401 application/problem+json`, not as a close code.
      tags: [momentums]
      security:
        - bearerAuth: []
      parameters:
        - name: from_height
          in: query
          required: false
          schema:
            type: integer
            format: int64
            minimum: 0
            examples: [12345000]
        - name: token
          in: query
          required: false
          description: |
            JWT fallback for browser clients. Identical semantics to
            the Authorization header — use one OR the other.
          schema:
            type: string
      responses:
        '101':
          description: Switching Protocols — WebSocket connection established.
        '400':
          description: Invalid `from_height` value.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Stream hub is not configured on this API instance.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/momentums/latest:
    get:
      operationId: getLatestMomentum
      summary: Get the most-recently-indexed momentum
      description: |
        Returns the highest-height momentum. Returns 404 if the indexer
        has not processed any blocks yet.
      tags: [momentums]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The latest momentum.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Momentum'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No momentums have been indexed yet.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/accounts/{address}:
    get:
      operationId: getAccount
      summary: Get account info
      description: |
        Returns the account's flow metrics (lifetime ZNN/QSR sent/received,
        first/last activity) plus delegation state. Returns 404 if the
        indexer has never observed the address.
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string, examples: ["z1qq…"] }
      responses:
        '200':
          description: Account info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No account with that address.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/accounts/{address}/balances:
    get:
      operationId: getAccountBalances
      summary: List balances for an account
      description: |
        Returns every (token_standard, balance) row for the address,
        sorted by balance DESC. Not paginated — an account typically
        holds a handful of tokens.
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: List of balances.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/accounts/{address}/transactions:
    get:
      operationId: listAccountTransactions
      summary: List transactions involving an address
      description: |
        Returns account-blocks where the address is either sender or
        recipient. Sorted by momentum_height; default desc.
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated list of account-blocks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountBlockList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/account_blocks:
    get:
      operationId: listAccountBlocks
      summary: List account-blocks (transactions)
      tags: [account_blocks]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated list of account-blocks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountBlockList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/account_blocks/{hash}:
    get:
      operationId: getAccountBlock
      summary: Get an account-block by hash
      tags: [account_blocks]
      security:
        - bearerAuth: []
      parameters:
        - name: hash
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The requested account-block.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountBlock'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No account-block with that hash.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/tokens:
    get:
      operationId: listTokens
      summary: List ZTS tokens
      description: |
        Returns ZTS tokens ordered by holder_count DESC (most-held first),
        then token_standard ASC for stability.
      tags: [tokens]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated list of tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/tokens/{token_standard}:
    get:
      operationId: getToken
      summary: Get a ZTS token by ZTS identifier
      tags: [tokens]
      security:
        - bearerAuth: []
      parameters:
        - name: token_standard
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Token info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No token with that standard.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/tokens/{token_standard}/holders:
    get:
      operationId: listTokenHolders
      summary: Richlist for a token
      description: |
        Returns addresses holding the token, sorted by balance DESC.
        Excludes zero balances.
      tags: [tokens]
      security:
        - bearerAuth: []
      parameters:
        - name: token_standard
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated holder list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalancePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/pillars:
    get:
      operationId: listPillars
      summary: List pillars
      description: |
        Returns pillars ordered by rank ascending. Excludes revoked
        pillars by default; pass `?include_revoked=true` to include them.
      tags: [pillars]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: include_revoked
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Paginated list of pillars.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PillarList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/pillars/{name}:
    get:
      operationId: getPillar
      summary: Get a pillar by name
      tags: [pillars]
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The requested pillar.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pillar'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No pillar with that name.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/pillars/{name}/delegators:
    get:
      operationId: listPillarDelegators
      summary: List delegators for a pillar
      description: |
        Returns the addresses currently delegated to the pillar, sorted
        by delegation_start_timestamp ASC (longest-tenured first).
      tags: [pillars]
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated delegator list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PillarDelegatorList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No pillar with that name.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/stakes:
    get:
      operationId: listStakes
      summary: List stake entries
      tags: [stakes]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: include_inactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Paginated stake list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StakeList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/accounts/{address}/stakes:
    get:
      operationId: listAccountStakes
      summary: List stakes owned by an address
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: include_inactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Paginated stake list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StakeList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/fusions:
    get:
      operationId: listFusions
      summary: List plasma fusion entries
      tags: [fusions]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: include_inactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Paginated fusion list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FusionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/accounts/{address}/fusions:
    get:
      operationId: listAccountFusions
      summary: List fusions for an address (funder or beneficiary)
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: include_inactive
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Paginated fusion list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FusionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/accounts/{address}/rewards:
    get:
      operationId: listAccountRewards
      summary: Per-event reward history for an address
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated reward-transaction list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RewardTransactionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/accounts/{address}/rewards/cumulative:
    get:
      operationId: getAccountCumulativeRewards
      summary: Cumulative reward totals for an address
      description: |
        Returns one row per (reward_type, token_standard) tuple — the
        lifetime sum the indexer has rolled up.
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cumulative reward totals.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CumulativeRewardList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/projects:
    get:
      operationId: listProjects
      summary: List Accelerator-Z projects
      tags: [projects]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated project list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/projects/{id}:
    get:
      operationId: getProject
      summary: Get a project by ID
      tags: [projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The requested project.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404':
          description: No project with that id.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/projects/{id}/phases:
    get:
      operationId: listProjectPhases
      summary: List phases of a project
      tags: [projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Phases in creation order.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectPhaseList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/projects/{id}/votes:
    get:
      operationId: listProjectVotes
      summary: List pillar votes targeting a project (or any of its phases)
      tags: [projects]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated vote list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VoteList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/bridge/wraps:
    get:
      operationId: listBridgeWraps
      summary: List wrap (Zenon → external chain) requests
      tags: [bridge]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated wrap-request list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WrapTokenRequestList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/bridge/unwraps:
    get:
      operationId: listBridgeUnwraps
      summary: List unwrap (external chain → Zenon) requests
      tags: [bridge]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated unwrap-request list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnwrapTokenRequestList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/accounts/{address}/bridge/wraps:
    get:
      operationId: listAccountBridgeWraps
      summary: List wraps with to_address = {address}
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated wrap-request list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WrapTokenRequestList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/v1/accounts/{address}/bridge/unwraps:
    get:
      operationId: listAccountBridgeUnwraps
      summary: List unwraps with to_address = {address}
      tags: [accounts]
      security:
        - bearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated unwrap-request list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnwrapTokenRequestList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...